55 multiple c sources (#56)
* more primitive types

- use all signed and unsigned primitive types
- allow all data types to have a default value. For primitive datatypes, a default value is required, so that a default value can always be generated for structs.

Note that you do not need "int" as a primitive type. (On all compilers that I know it's the same as long.)

* fix threadpool

zie het oude commentaar voor wat het probleem was

* beginnetje

* moved task-related functionality to different package for reuse

* first working version of multiple c files

The producer works. It creates source files. Then it compiles them to bitcode (LLVM IR), human-readable LLVM IR and an executable with every compiler configuration. Also: some task classes are moved out of devtools.explorer, and there are some enhancements to the explorer that I needed.

* bug fixed, improvements

- bug fixed in ProcessTask
- A second c-file is used, by the DataStructureFeature
- added a little bit of API to CGenerator

---------

Co-authored-by: dinkydau <onzin@onzin.nl>