complete recursive parsing of type specifiers
For structs, for example, that means all members of the struct are also parsed, and if any of those is a struct, the members of that struct are parsed etc.

It works but needs improvement in the handling of errors. Also the normalizeCode functions need to be reconsidered. How to bring the C code into a form that is easier to parse manually (for example, the only whitespace is spaces, and there are no comments) and when should that be done?