47 datastructure feature (#76)
* more primitive types

- use all signed and unsigned primitive types
- allow all data types to have a default value. For primitive datatypes, a default value is required, so that a default value can always be generated for structs.

Note that you do not need "int" as a primitive type. (On all compilers that I know it's the same as long.)

* fix threadpool

zie het oude commentaar voor wat het probleem was

* beginnetje

* moved task-related functionality to different package for reuse

* first working version of multiple c files

The producer works. It creates source files. Then it compiles them to bitcode (LLVM IR), human-readable LLVM IR and an executable with every compiler configuration. Also: some task classes are moved out of devtools.explorer, and there are some enhancements to the explorer that I needed.

* work in progress

* work in progress: assessor keeps decompiled code

* progress

* extraction of testcases from decompiled C

It seems to be working but is definitely not finished. This is a first step.

* next step: generalize VariableInfo classes to NameInfo classes

* backup

* refactor DataStructureCVisitor

* all (or almost all, because there will always be exceptions) declarations can be converted to an internal representation

* prevent lawsuits

* make parsing of complex declarators possible

* proper handling of typedefs

* complete recursive parsing of type specifiers

For structs, for example, that means all members of the struct are also parsed, and if any of those is a struct, the members of that struct are parsed etc.

It works but needs improvement in the handling of errors. Also the normalizeCode functions need to be reconsidered. How to bring the C code into a form that is easier to parse manually (for example, the only whitespace is spaces, and there are no comments) and when should that be done?

* recovering testcases from decompiled code should work

* Update Parsing.java

* state before merging of master

* there is an index out of bounds exception in the producer

* codemarker extensions needed for the datastructure feature

* make it possible to check if a codemarker was correctly parsed from a string

* Update Main.java

* Update CGenerator.java

* before retry of merge codemarker branch

* remove old version of DataStructuresFeature

* the producer works

* bugs in the declarator parser

* enum test

* first version that does something

* generated C code seems to always compile

* producer can be considered done

* before merge

---------

Co-authored-by: dinkydau <onzin@onzin.nl>
Co-authored-by: reijer-dev <76157894+reijer-dev@users.noreply.github.com>